Fix catalog initialization and loading tools

+++ leads-aux-tools/load-amplab/src/main/java/data/LoadAmplab.java
@@ -362,7 +362,7 @@ public class LoadAmplab {
                                 else if (pair[1].toLowerCase().equals("bigint"))
                                     columnType.add(Long.class);
                                 else if (pair[1].toLowerCase().equals("int"))
-                                    columnType.add(Long.class);
+                                    columnType.add(Integer.class);
                                 else if (pair[1].toLowerCase().equals("float"))
                                     columnType.add(Float.class);
                                 else {
